add bytestrings, rename define-core to define, support top-interaction
Strings are now compiled to plutus bytestrings. This is implemented with
`string->bytes/locale', which is probably not ideal, but it works fine.

`define' is preferable to `define-core' because it exhibits less
syntactic noise.

`#%top-interaction' now behaves as expected and interacting with core in
the repl is much more pleasant.